+++ src/FFI.php
@@ -178,6 +178,18 @@ class FFI
         self::vips()->vips_shutdown();
     }
 
+    public static function newGClosure(): \FFI\CData
+    {
+        // GClosure measures 32-bit with the first few fields until marshal
+        // Marshal is a function pointer, thus platform-dependant.
+        // Data is a pointer, thus platform-dependant.
+        // Notifiers is an array-pointer, thus platform-dependant.
+        // All in all it's basically 4 (bytes) + 3 * POINTER_SIZE
+        // However, gobject wants 8 (bytes) + 3 * POINTER_SIZE.
+        // I'm not sure where that extra byte comes from. Padding on 64-bit machines?
+        return self::gobject()->g_closure_new_simple(8 + 3 * PHP_INT_SIZE, null);
+    }
+
     private static function libraryName(string $name, int $abi): string
     {
         switch (PHP_OS_FAMILY) {
@@ -191,6 +203,29 @@ class FFI
             default:
                 // most *nix
                 return "$name.so.$abi";
+        }
+
+        return null;
+    }
+
+    private static function libraryLoad(
+        array $libraryPaths,
+        string $libraryName,
+        string $interface
+    ): \FFI {
+        Utils::debugLog("trying to open", ["libraryName" => $libraryName]);
+        foreach ($libraryPaths as $path) {
+            Utils::debugLog("trying path", ["path" => $path]);
+            try {
+                $library = \FFI::cdef($interface, $path . $libraryName);
+                Utils::debugLog("success", []);
+                return $library;
+            } catch (\FFI\Exception $e) {
+                Utils::debugLog("init", [
+                    "msg" => "library load failed",
+                    "exception" => $e->getMessage()
+                ]);
+            }
         }
     }
 
@@ -239,30 +274,15 @@ class FFI
             $libraryPaths[] = "/opt/homebrew/lib/"; // Homebrew on Apple Silicon
         }
 
-        // attempt to open libraries using the system library search method
-        // (no prefix) and a couple of fallback paths, if any
-        $vips = null;
-        foreach ($libraryPaths as $path) {
-            Utils::debugLog("init", ["path" => $path]);
-
-            try {
-                $vips = \FFI::cdef(<<<EOS
-                    int vips_init (const char *argv0);
-                    const char *vips_error_buffer (void);
-                    int vips_version(int flag);
-                EOS, $path . $vips_libname);
-                break;
-            } catch (\FFI\Exception $e) {
-                Utils::debugLog("init", [
-                    "msg" => "library load failed", 
-                    "exception" => $e->getMessage()
-                ]);
-            }
-        }
+        $vips = self::libraryLoad($libraryPaths, $vips_libname, <<<EOS
+            int vips_init (const char *argv0);
+            const char *vips_error_buffer (void);
+            int vips_version(int flag);
+        EOS);
 
         if ($vips === null) {
+            // drop the "" (system path) member
             array_shift($libraryPaths);
-
             $msg = "Unable to open library '$vips_libname'";
             if (!empty($libraryPaths)) {
                 $msg .= " in any of ['" . implode("', '", $libraryPaths) . "']";
@@ -305,6 +325,7 @@ typedef uint32_t guint32;
 typedef int32_t gint32;
 typedef uint64_t guint64;
 typedef int64_t gint64;
+typedef void* gpointer;
 
 typedef $gtype GType;
 
@@ -369,6 +390,7 @@ void g_value_set_enum (GValue* value, int e);
 void g_value_set_flags (GValue* value, unsigned int f);
 void g_value_set_string (GValue* value, const char* str);
 void g_value_set_object (GValue* value, void* object);
+void g_value_set_pointer (GValue* value, gpointer pointer);
 
 bool g_value_get_boolean (const GValue* value);
 int g_value_get_int (GValue* value);
@@ -379,6 +401,7 @@ int g_value_get_enum (GValue* value);
 unsigned int g_value_get_flags (GValue* value);
 const char* g_value_get_string (GValue* value);
 void* g_value_get_object (GValue* value);
+gpointer g_value_get_pointer (GValue* value);
 
 typedef struct _GEnumValue {
     int value;
@@ -432,6 +455,19 @@ long g_signal_connect_data (GObject* object,
     int connect_flags);
 
 const char* g_param_spec_get_blurb (GParamSpec* psp);
+
+typedef void *GClosure;
+typedef void (*marshaler)(
+    struct GClosure* closure, 
+    GValue* return_value, 
+    int n_param_values, 
+    const GValue* param_values, 
+    void* invocation_hint, 
+    void* marshal_data
+);
+void g_closure_set_marshal(GClosure* closure, marshaler marshal);
+long g_signal_connect_closure(GObject* object, const char* detailed_signal, GClosure *closure, bool after);
+GClosure* g_closure_new_simple (int sizeof_closure, void* data);
 EOS;
 
         # the whole libvips API, mostly adapted from pyvips
@@ -551,7 +587,7 @@ typedef struct _VipsArgumentClass {
     unsigned int offset;
 } VipsArgumentClass;
 
-int vips_object_get_argument (VipsObject* object, const char *name, 
+int vips_object_get_argument (VipsObject* object, const char *name,
     GParamSpec** pspec,
     VipsArgumentClass** argument_class,
     VipsArgumentInstance** argument_instance);
@@ -706,12 +742,6 @@ typedef struct _VipsSourceCustom {
 
 VipsSourceCustom* vips_source_custom_new (void);
 
-// FIXME ... these need porting to php-ffi
-// extern "Python" gint64 _marshal_read (VipsSource*,
-//    void*, gint64, void*);
-// extern "Python" gint64 _marshal_seek (VipsSource*,
-//    gint64, int, void*);
-
 typedef struct _VipsTarget {
     VipsConnection parent_object;
 
@@ -736,9 +766,21 @@ EOS;
         }
 
         Utils::debugLog("init", ["binding ..."]);
-        self::$glib = \FFI::cdef($glib_decls, $glib_libname);
-        self::$gobject = \FFI::cdef($gobject_decls, $gobject_libname);
-        self::$vips = \FFI::cdef($vips_decls, $vips_libname);
+        self::$glib = self::libraryLoad(
+            $libraryPaths,
+            $glib_libname,
+            $glib_decls
+        );
+        self::$gobject = self::libraryLoad(
+            $libraryPaths,
+            $gobject_libname,
+            $gobject_decls
+        );
+        self::$vips = self::libraryLoad(
+            $libraryPaths,
+            $vips_libname,
+            $vips_decls
+        );
 
         # Useful for debugging
         # self::$vips->vips_leak_set(1);
@@ -755,11 +797,18 @@ EOS;
         // look these up in advance
         self::$ctypes = [
             "GObject" => self::$gobject->type("GObject*"),
+            "GClosure" => self::$gobject->type("GClosure"),
             "GParamSpec" => self::$gobject->type("GParamSpec*"),
             "VipsObject" => self::$vips->type("VipsObject*"),
             "VipsOperation" => self::$vips->type("VipsOperation*"),
             "VipsImage" => self::$vips->type("VipsImage*"),
             "VipsInterpolate" => self::$vips->type("VipsInterpolate*"),
+            "VipsConnection" => self::$vips->type("VipsConnection*"),
+            "VipsSource" => self::$vips->type("VipsSource*"),
+            "VipsSourceCustom" => self::$vips->type("VipsSourceCustom*"),
+            "VipsTarget" => self::$vips->type("VipsTarget*"),
+            "VipsTargetCustom" => self::$vips->type("VipsTargetCustom*"),
+            "VipsProgress" => self::$vips->type("VipsProgress*"),
         ];
 
         self::$gtypes = [
@@ -783,6 +832,8 @@ EOS;
 
             "GObject" => self::$gobject->g_type_from_name("GObject"),
             "VipsImage" => self::$gobject->g_type_from_name("VipsImage"),
+
+            "GClosure" => self::$gobject->g_type_from_name("GClosure"),
         ];
 
         // map vips format names to c type names