+++ src/TextWrap.php
@@ -39,7 +39,7 @@
 namespace Jcupitt\Vips;
 
 /**
- * The ForeignPpmFormat enum.
+ * The TextWrap enum.
  * @category  Images
  * @package   Jcupitt\Vips
  * @author    John Cupitt <jcupitt@gmail.com>
@@ -47,11 +47,10 @@ namespace Jcupitt\Vips;
  * @license   https://opensource.org/licenses/MIT MIT
  * @link      https://github.com/jcupitt/php-vips
  */
-abstract class ForeignPpmFormat
+abstract class TextWrap
 {
-    const PBM = 'pbm';
-    const PGM = 'pgm';
-    const PPM = 'ppm';
-    const PFM = 'pfm';
-    const PNM = 'pnm';
+    const WORD = 'word';
+    const CHAR = 'char';
+    const WORD_CHAR = 'word-char';
+    const NONE = 'none';
 }